refactor(bayn): consolidate execution kernel
