+++ services/bayn/src/paper-candidate-discovery/model.ts
@@ -14,7 +14,8 @@ import {
 import { RuntimeProvenanceSchema } from '../contracts'
 import type { AutonomousCycle } from '../cycle'
 import type { CycleOperationsProjection } from '../cycle-observability'
-import { Authority, OrderSide, OrderType, TimeInForce } from '../paper'
+import { OrderSide, OrderType, TimeInForce } from '../execution/contracts'
+import { Authority } from '../execution/legacy-paper-codecs'
 import {
   GitSourceRevisionSchema,
   ImageDigestSchema,
