+++ services/bayn/src/paper-candidate-discovery/program.ts
@@ -6,26 +6,14 @@ import { CycleObservability, type CycleObservabilityError } from '../db/cycle-ob
 import { CycleStore, type CycleStoreError } from '../db/cycle-store'
 import type { AutonomousCycle } from '../cycle'
 import type { ObserveShadowDecisionDocument } from '../shadow-decision-contract'
-import {
-  assetReadConcurrency,
-  type PaperCandidateDiscoveryIdentity,
-  type PaperCandidateDiscoveryReceipt,
-  type PaperCandidateDiscoverySnapshot,
-  type ValidatedAccount,
-  type ValidatedAccountConfiguration,
-  type ValidatedAssets,
-  type ValidatedPaperCandidateObservations,
-  type ValidatedPaperCandidateSnapshot,
+import type {
+  PaperCandidateDiscoveryIdentity,
+  PaperCandidateDiscoveryReceipt,
+  PaperCandidateDiscoverySnapshot,
 } from './model'
 import { isPaperCandidateDiscoveryError, requireValue, type PaperCandidateDiscoveryError } from './failure'
-import { selectCompletedCycle, validateIdentity, validateSnapshotForIdentity } from './snapshot-validation'
-import {
-  assembleValidatedObservations,
-  validateAccountConfiguration,
-  validateAccountObservation,
-  validateAssetObservations,
-} from './broker-observation-validation'
-import { makePaperCandidateDiscoveryReceipt } from './receipt-construction'
+import { selectCompletedCycle } from './snapshot-validation'
+import { makeCandidateDiscovery } from './interpreter'
 
 const readCycle = (
   store: CycleStore['Service'],
@@ -113,87 +101,6 @@ const readDiscoverySnapshot = (
     ),
   )
 
-const readAccount = (
-  broker: BrokerRead['Service'],
-  identity: PaperCandidateDiscoveryIdentity,
-): Effect.Effect<ValidatedAccount, PaperCandidateDiscoveryError> =>
-  pipe(
-    broker.account,
-    Effect.mapError(
-      (cause): PaperCandidateDiscoveryError => ({
-        _tag: 'BrokerReadFailed',
-        failure: 'broker',
-        read: 'account',
-        accountId: identity.accountId,
-        cause,
-      }),
-    ),
-    Effect.flatMap((account) => Effect.fromResult(validateAccountObservation(identity, account))),
-  )
-
-const readAccountConfiguration = (
-  broker: BrokerRead['Service'],
-  account: ValidatedAccount,
-): Effect.Effect<ValidatedAccountConfiguration, PaperCandidateDiscoveryError> =>
-  pipe(
-    broker.accountConfiguration,
-    Effect.mapError(
-      (cause): PaperCandidateDiscoveryError => ({
-        _tag: 'BrokerReadFailed',
-        failure: 'broker',
-        read: 'account-configuration',
-        accountId: account.read.value.id,
-        cause,
-      }),
-    ),
-    Effect.flatMap((configuration) => Effect.fromResult(validateAccountConfiguration(account, configuration))),
-  )
-
-const readAssets = (
-  broker: BrokerRead['Service'],
-  snapshot: PaperCandidateDiscoverySnapshot,
-  configuration: ValidatedAccountConfiguration,
-): Effect.Effect<ValidatedAssets, PaperCandidateDiscoveryError> =>
-  pipe(
-    Effect.forEach(snapshot.document.targetPlan.intentTargets, (intent) => broker.assetBySymbol(intent.symbol), {
-      concurrency: assetReadConcurrency,
-    }),
-    Effect.mapError(
-      (cause): PaperCandidateDiscoveryError => ({
-        _tag: 'BrokerReadFailed',
-        failure: 'broker',
-        read: 'assets',
-        accountId: snapshot.cycle.identity.accountId,
-        symbols: snapshot.document.targetPlan.intentTargets.map(({ symbol }) => symbol),
-        cause,
-      }),
-    ),
-    Effect.flatMap((assets) => Effect.fromResult(validateAssetObservations(snapshot, configuration, assets))),
-  )
-
-const observeBroker = (
-  validatedSnapshot: ValidatedPaperCandidateSnapshot,
-): Effect.Effect<ValidatedPaperCandidateObservations, PaperCandidateDiscoveryError, BrokerRead> =>
-  pipe(
-    BrokerRead,
-    Effect.flatMap((broker) =>
-      pipe(
-        Effect.Do,
-        Effect.bind('account', () => readAccount(broker, validatedSnapshot.identity)),
-        Effect.bind('accountConfiguration', ({ account }) => readAccountConfiguration(broker, account)),
-        Effect.bind('assets', ({ accountConfiguration }) =>
-          readAssets(broker, validatedSnapshot.snapshot, accountConfiguration),
-        ),
-        Effect.bind('capturedAtMs', () => Clock.currentTimeMillis),
-        Effect.flatMap(({ account, accountConfiguration, assets, capturedAtMs }) =>
-          Effect.fromResult(
-            assembleValidatedObservations(validatedSnapshot, account, accountConfiguration, assets, capturedAtMs),
-          ),
-        ),
-      ),
-    ),
-  )
-
 export const discoverPaperCandidates = (
   candidateIdentity: PaperCandidateDiscoveryIdentity,
 ): Effect.Effect<
@@ -201,21 +108,13 @@ export const discoverPaperCandidates = (
   PaperCandidateDiscoveryError,
   PgClient.PgClient | CycleObservability | CycleStore | BrokerRead
 > =>
-  pipe(
-    validateIdentity(candidateIdentity),
-    Effect.fromResult,
-    Effect.flatMap((identity) =>
-      pipe(
-        Effect.Do,
-        Effect.bind('snapshot', () => readDiscoverySnapshot(identity)),
-        Effect.bind('startedAt', () => Clock.currentTimeMillis),
-        Effect.bind('validatedSnapshot', ({ snapshot, startedAt }) =>
-          Effect.fromResult(validateSnapshotForIdentity(identity, snapshot, startedAt)),
-        ),
-        Effect.bind('observations', ({ validatedSnapshot }) => observeBroker(validatedSnapshot)),
-        Effect.flatMap(({ observations, validatedSnapshot }) =>
-          Effect.fromResult(makePaperCandidateDiscoveryReceipt(validatedSnapshot, observations)),
-        ),
-      ),
-    ),
-  )
+  Effect.gen(function* () {
+    const broker = yield* BrokerRead
+    return yield* makeCandidateDiscovery({
+      readSnapshot: readDiscoverySnapshot,
+      readAccount: broker.account,
+      readAccountConfiguration: broker.accountConfiguration,
+      readAssetBySymbol: broker.assetBySymbol,
+      currentTimeMillis: Clock.currentTimeMillis,
+    }).discover(candidateIdentity)
+  })
