+++ services/bayn/src/paper-candidate-discovery/snapshot-validation.ts
@@ -3,7 +3,8 @@ import { Result, Schema, pipe } from 'effect'
 import { makeStrategyProtocolHash } from '../contracts'
 import { CycleState } from '../cycle'
 import type { CycleOperationsProjection } from '../cycle-observability'
-import { Authority, RiskOutcome } from '../paper'
+import { RiskOutcome } from '../execution/contracts'
+import { Authority } from '../execution/legacy-paper-codecs'
 import { Gate, Reason } from '../risk'
 import { strictParseOptions } from '../schemas'
 import type { ObserveShadowDecisionDocument } from '../shadow-decision-contract'
